# src/main/java/com/example/bankcards/dto/TransferRequest.java
package com.example.bankcards.dto;

import lombok.Data;

import java.math.BigDecimal;

@Data
public class TransferRequest {
    private Long fromCardId;
    private Long toCardId;
    private BigDecimal amount;
}